Math: Add Vec4ui

+++ Src/Engine/Common/Math/Simd8Ray.hpp
@@ -49,9 +49,9 @@ public:
     // otherwise, returns 0xFFFFFFFF
     NFE_FORCE_INLINE uint32 GetOctant() const
     {
-        const size_t countX = Common::BitUtils<uint32>::CountBits(dir.x.GetSignMask());
-        const size_t countY = Common::BitUtils<uint32>::CountBits(dir.y.GetSignMask());
-        const size_t countZ = Common::BitUtils<uint32>::CountBits(dir.z.GetSignMask());
+        const uint32 countX = Common::BitUtils<uint32>::CountBits(dir.x.GetSignMask());
+        const uint32 countY = Common::BitUtils<uint32>::CountBits(dir.y.GetSignMask());
+        const uint32 countZ = Common::BitUtils<uint32>::CountBits(dir.z.GetSignMask());
 
         const uint32 xPart[9] = { 0u, ~0u, ~0u, ~0u, ~0u, ~0u, ~0u, ~0u, 1u << 0u };
         const uint32 yPart[9] = { 0u, ~0u, ~0u, ~0u, ~0u, ~0u, ~0u, ~0u, 1u << 1u };
